+++ ddscatcli/__init__.py
@@ -1,1 +1,3 @@
-__version__ = "1.0.0"
+__version__ = "1.0.2"
+
+from .ddscat_cli import run  # noqa: F401